bitcoin_monitor/requirements.txt
aiohttp
aiofiles
aiosqlite
aiosmtplib
tortoise-orm
schedule
python-dotenv